Added foreign keys asociation to Order, User and Payment models, order controller: index and store are working

+++ app/Http/Requests/StoreBookRequest.php
@@ -26,7 +26,7 @@ class StoreBookRequest extends FormRequest
         return [
             'publisher' => ['required','string'],
             'title' => ['required','string'],
-            'price' => ['required','integer'],
+            'price' => ['required','numeric'],
             'publish_year'=> ['required','integer'],
             'picture' => ['required', 'image'],
             'author_ids' => ['required','array'],
